feat(library): Introduce SQLite
Introduce SQLite native application for Unikraft.

Add required files: `Kraftfile`, `Makefile.uk` and `README.md`.

The root filesystem (in `rootfs/`) provides a database file (`music.db`)
that is to be queried with `query.sql`. The querying is defined as the
`cmd` in `Kraftfile`. The `build.sql` script is used to generate the
`music.db` database file.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>