fix(merge-oci-digests): Show `--all` packages
Following the merge of kraftkit#945, without specifying the `--all`
flag, the host architecture and platform are detected and used to reduce
the output.  In this CI/CD context, we wish to show all variants which
requires the use of this flag.  Without this flag, the GitHub Actions
workflow fails as it cannot determine the host platform.

Signed-off-by: Alexander Jung <alex@unikraft.io>