refactor(library/base): Introduce fallback filesystem
This commit introduces a fallback filesystem to the base image
following the revamp of the automount code in Unikraft[0].

The fallback filesystem contains a single file with a helpful
message informing the user that the base image has not been
supplied with a filesystem.  Upon providing a filesystem to
the base image, the embedded fallback filesystem is not
mounted and the user's code can be executed.

[0]: https://github.com/unikraft/unikraft/pull/1262

Signed-off-by: Alexander Jung <alex@unikraft.io>