fix(merge-oci-digests): Downgrade `actions/download-artifact`
This is because the accompanying action, `actions/upload-artifact`
is still at v3 which is stated as non-compatible.

Additionally, the README for v4 indicates major changes and
backward compatibility issues.  The easy fix for now to
solve merging OCI manifests together (which fails currently
because it cannot retrieve the downloaded artifacts from the
previous dependent workflows), is just to simply downgrade.

A solution in the future which upgrades must also adjust the
action settings in order to achieve the same functionality.

Signed-off-by: Alexander Jung <alex@unikraft.io>