fix(base): Remove `CONFIG_LWIP_FAILNOIFACE`
The option `CONFIG_LWIP_FAILNOIFACE` makes sense to enable only for
applications that rely on networking. Where networking is optional, this
KConfig option should be set to `n` and where no networking is needed, we
ideally want to build a Unikernel without networking capabilities and
drivers. The `base` image is a generic elfloader image where we are
weakening the specialization degree because we need to think a bit
more "general purpose": Applications without networking requirements
should be able to be run as well.

Github-Fixes: #135
Signed-off-by: Simon Kuenzer <simon@unikraft.io>