fix(library): Add tty stdio configuration
Unikraft PR #1437 introduced the `LIBPOSIX_TTY_STDIO` config option.
This must be enabled in order to use standard descriptors. In
particular, this enables applications to print to standard output. If
the option is not enabled, nothing is printed. It seems as if the
application is doing nothing or hangs (in case of servers).

Enable the `LIBPOSIX_TTY_STDIO` in `Kraftfile`s for native and
binary-compatibility library builds.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>