fix(library): Fix Dockerfile for PHP 8.2
Use correct library names in Dockerfile. The previous versions of the
library were missing. Not sure if it was caused by an upstream update of
the `php8.2-cli` Docker image, or they were faulty from the start.

Also use a single stage to copy all libraries in `/lib/x86_64-linux-gnu`
and sort library files alphabetically.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>